+++ backend/config/dotenv.js
@@ -1,0 +1,2 @@
+const dotenv = require('dotenv');
+dotenv.config();
