+++ backend/middleware/errorHandler.js
@@ -1,0 +1,6 @@
+function errorHandler(err, req, res, next) {
+  console.error(err.stack);
+  res.status(500).send('Something went wrong');
+}
+
+module.exports = errorHandler;
